Add files via upload
Median house prices for California districts  from the 1990 census. This is a modified version of the dataset on https://towardsdatascience.com/26-datasets-for-your-data-science-projects-658601590a4c . Certain fields have been taken away such as longitude and latitude. The dataset was used in the second chapter of Aurélien Géron's  book 'Hands-On Machine learning with Scikit-Learn and TensorFlow' and was noted as such in the aforementioned website.